requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
Faker==13.14.0
idna==3.3
mimesis==5.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
urllib3==1.26.9
